requirements.txt
opencv-python
Pillow
numpy
pyzmq
python-telegram-bot
pika